Create TECHNOLOGY
TECHNOLOGY CHANGING THE FACE OF EDUCATION?
HomeTechnology changing the face of education?

 
Spread the love
By Rajdeep kaur

Technology changing the face of education: Technology cannot be terminated without the utilization of procedures, tools, machines, materials, and sources of power to make human life more comfortable throughout. it has aided in changing the face of education and developing the education system and productivity as well.


technology advancement
It encourages one’s abilities and permits one to use the immense material forces of creation.  Technological advancement is the driving force of commercial growth.  Consequently, the availability of technology defines the efficiency of stock and the acceleration of economic growth.

Technology is essential in our life
It has the main purpose of solving the query of humankind and making work more peaceful. We have been massively dependent on technology that we cannot do anything without it. Technology is used everywhere when we utilize mobile phones, drive a car, watch tv, electrical engines and use the computer, etc . with the maintenance of technology, we can bright not only the present but also the future.

Technology development in several areas
In the earlier years, the masses have made excellent advancements in technology. Present-day India has had a robust focus on technology and science that designates it as a core component for economic growth and cultural development. Development is progressing in the various fields i.e. product and service of information of technology, the building of electronics and transportation, etc.

Technology boosts in artificial intelligence, robotic process automation, edge computing, virtual reality, cybersecurity, blockchain, quantum computing, internet of things, 5g.

The technology department gives goods and services that are used in consumer and merchandising functions. Consumer commodities are mobiles, wearable technology, computers and home appliances. Business uses technology to handle their databases, preserve financial data, performances of business, and overall outcomes.

Technology transforming the appearance of teaching

technology transforming the appearance of teaching
Technology has converted an enormous part of society and day-to-day life. it is utilized to assist both teaching and learning technology in the classroom.  technology is a predominant mechanism that supports and moulds education. it has the capability to enhance connections between teachers and students.

It can put all the data into one position, promote learning motivation, it also intensifies learning efficiency because of visualization of the thought.  technology presents a satisfactory environment in which students discover and solve their own troubles. Students are also capable to cooperate with their own classroom through technology purposes.

we have multiple platforms on which we can communicate with friends, and teachers. those platforms are Google meet, Skype, Slack, zoom, and Bigbluebutton, etc. Zoom is one of the most prevalent conferencing platforms.

Technology has a positive and a negative influence.
No doubt, technology has enhanced a significant factor of daily life. but it has a positive impact and a negative consequence.

POSITIVE EFFECT
1 we can reach distant places within hours.

2 save time and money

3 enhancement of learning skills

4 development of social interaction

5 Problem-solving abilities

6 with the help of mobile technology,  we can easily reach our friends and relatives who are living far from us

7 enhancement of student engagement

8 through the internet, we are able to learn new things and online lectures

NEGATIVE EFFECT
1 Due to the evolution of technology workers face difficulties means it generates unemployment.

2 It is also created depressing social issues, health problems.

3 reduce slumber due to mobile devices.

4 less physical motions

5 feeling lonely

6 neck pain and bad posture.

7 kids play those games which are dangerous in their life.

8 somebody utilizes technology in the wrong way which destroys their life. an eg. terrorist, thieves.

CONCLUSION
We can observe nowadays, without technology we are worthless because everyone wants to use technology to facilitate work. with technology, teachers have discovered new techniques to interact with their pupils. so we can say technology changing the face of education. it improves interaction, learning system, resourcing sharing by innovative audio-video characteristics. it helps administration, businesses, and people increase their efficiency and effectiveness.